Add "declare module" wrapper to declaration file
Eventually this should be something set by a flag, since there is a use
case for haing and not having it.

Users would want `declare module "*.elm"` when they're using webpack or
some other asset manager for Elm, and want to `import {Elm} from "Foo.elm"`
in their typescript code.

Users would not want the wrapper when they are using some other method
to compile Elm, and then using the js with `import {Elm} from "Foo"`.